Rewrite pending chip positioning algorithm
Originally, the pending chip would follow the cursor only if the cursor
was over a grid column, a chip slot, or a placed chip. However, when
placing chips on the grid, the #chips container would sometimes get
between the cursor and a grid column / chip slot, thereby skewing the
position calculations because the event.target referred to the container
and not a grid column or chip slot as expected.