Eliminate obsolete current translate value logic
At one point, I was calling setTranslate with either an x or a y
coordinate (but not both), and so I added logic to retrieve the current
translate value for whichever coordinate was not supplied explicitly.
However, every one of my calls to setTranslate now supplies both
coordinates, making the aforementioned code no longer necessary.