Remove 1-player mode from app
Implementing 1-player mode (which implies the presence of an AI player)
would require a great deal of effort, mostly re: how the AI interacts
with the game in the same way a human player would.

The app assumes that any player uses their cursor to place each chip
on the grid. However, an AI player would not have a cursor, so it would
need to modify the grid programatically and also trigger the chip drop
animation. Plus, the intelligence of the AI itself must be programmed,
which would require extensive research and thought. Therefore, for all
of these reasons, adding a 1-Player mode is no longer a goal for v1.0