Fix bug where pending chip position would not reset
This bug occurred primarily on touch devices, where if the the user
initially tapped the leftmost column (where the pending chip is
initially positioned), then the chip would be placed as normal, but the
pending chip's position would not reset to above the pointer column.
This is because the pointerColumnX property was only set on mousemove,
and not on click.