Move general chip placing logic to separate function
This commit moves the event-agnostic of placing the pending chip from
the event handler to a separate function. This will allow for the
pending chip to be placed programatically (with transitions and all),
which may ease the task of implementing an AI in the future.

Please note the event-agnostic logic in movePendingChipViaPointer still
needs to be abstracted out, so for the time being, gameplay on touch
devices is broken.