Move connection-capping logic to Game model
When a connection of more than four chips has been found, only some
group of four chips within that connection is actually highlighted.
Previously, this connection-capping logic resided in
Grid.getConnections, however having the logic there limits the usability
of the function (e.g. what if you want to examine all chips in the
connection and not just four?). Therefore, the capping logic has been
moved from the Grid model to the Game model where it ought to reside.