Remove redundant startGame() calls in AI tests
None of the AI player tests actually need to start the game, because the
chips are placed on the grid directly rather than through the game
object. This is done so that the test case's column list can bypass the
automatic computing of the AI when its turn starts (this ensuring that
future revisions of the AI don't require changing these column lists).