Correct isMaxPlayer to currentPlayerIsMaxPlayer
This commit fixes an obvious typo, but one which actually did not
produce any incorrect behavior. This is due to the coincidental fact
that `undefined` has the same "falsiness" as `false`, and since the
checks within getScore() do not check for explicit values (e.g. `===
false`), then the typo doesn't actually break any functionality.