Remove no-restricted-syntax rule from eslintrc
The no-restricted-syntax rule doesn't do much for features that I never
use anyway (especially those features which are universally considered
bad practice, like the `with` statement). The usage of the rule is also
more complicated, as one needs to look up a list of ESTree node types in
order to set the correct values.