Eliminate aiAutoPlace parameter
The aiAutoPlace parameter was originally added to instruct the
movePendingChipToColumn function to automatically place the AI player's
pending chip after aligning it. However, the addition of this property
reflects a poor separation of concerns: the movePendingChipToColumn
function should not be concerned with the placing of the chip; such
logic belongs in the placePendingChip function.

To resolve this issue of separation, allow the movePendingChipToColumn
function to accept an optional callback to run when the pending chip
transition ends. The placePendingChip function will then utilize this
optional callback to automatically place the AI player's chip when the
pending chip alignment has finished. In effect, the task of placing the
AI's chip becomes the responsibility of placePendingChip, as it should
always have been.