Quote keys in Sass map of player colors
This solves a strange issue where the AI player's chips did not contain
a black fill. The fix involved quoting the unquoted color names
used as the keys for the player color map in Sass. The reasons for why
this works are largely unknown, but I suspect the color names were
somehow being interpreted as colors rather than map keys. Regardless,
this fix proves the wisdom in always quoting one's key names.