Fix AI's handling of double-win scenarios
This commit definitively fixes the AI's handling of double-win
scenario. That is, the AI will now always block an opponent win, even if
the opponent is guaranteed to win on its next move. The two
recently-added tests have also been uncommented and are both passing
along with all other tests. The third diagonal win test was also
reclassified as a double-win test after further examination.