Abandon Vitest migration; reverting to Playwright
This is because many of the UI tests are testing placing chips with a
physical mouse/pointer, which in itself requires a browser layout engine
(e.g. for event coordinates and offsets and such). There is no way to
work around this in Vitest without exposing "shortcuts" in the
application code to make testing easier (e.g. programmatically simulate
placing a chip when a pointing device isn't available). That would be a
rather hacky workaround, though, and one which disregards how the user
will actually interact with the application.